Poprawki dla webserviceslib: 1) Poprawiono obsługe błędów serwisów webowych 2) Poprawka dla generowania czasu na potrzeby generowania zapytań do serwisów 3) Poprawka dla parsowania płytkiej kopii danych - wyciąganie daty diagnozy, może być puste 4) Zmiana reprezentacji daty urodzenia pacjenta po stronie płytkiej kopii danych
git-svn-id: https://localhost:8443/svn/hm-mde-1@974 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edrutils/src/webserviceslib/src/DateTime.cpp
@@ -75,9 +75,9 @@ const std::string Date::toString() const
 
     sstream.flags(std::ios::right);
     sstream << year << "-";
-    sstream.width(2); sstream << month;
+    sstream.width(2); sstream << (unsigned int)month;
     sstream << "-";
-    sstream.width(2); sstream << day;
+    sstream.width(2); sstream << (unsigned int)day;
 
     return sstream.str();
 }
@@ -237,10 +237,10 @@ const std::string Time::toString() const
     sstream.fill('0');
     sstream.width(2);
     sstream.flags(std::ios::right);
-    sstream << hour << ":";
-    sstream.width(2); sstream << minutes;
+    sstream << (unsigned int)hour << ":";
+    sstream.width(2); sstream << (unsigned int)minutes;
     sstream << ":";
-    sstream.width(2); sstream << seconds;
+    sstream.width(2); sstream << (unsigned int)seconds;
 
     return sstream.str();
 }
